Add files via upload
Updated the code to respond to POST requests at endpoint /api/spells/8, and DELETE requests at the same by adding a spell, and deleting the same spell respectively. Also added a PUT request that fixes a misspelling of one of the database entries. 
TODO:
Add a revert function that reverts Black Tenticles to the misspelled state, so that it can be run more than once without updating the database manually.
Fix various things, including a vulnerability to SQL injection (I think it is, anyway). (Since this isn't production code, I did that long line of SQL that way, but normally, I wouldn't list data in the SQL, and would just list the data as "?", and use an array.)